updated collision detection to specify specific tile constraints
This update allows for object data to control the exact collision constraint of each side of a tile. for example if you have "w":[10,10,5,10]

the constraint is tells the collision function the amount of space the sprite can walk past the x y boundary. the values represent North, East, South, West.

This change is useful for objects that are not completely solid for example palm trees. some tiles have branches or other attributes that have extra space to walk into.