simplified room data and changed naming convention
I have removed the layout and index attributes from the items.json and instead integrated it into the actual data. I have also changed the naming convention between the items dictionary and the room data as they shared the same x y but meant different things. the x and y in room data is the position on the scene map and the new a & b in the dictionary is the top and left offset within the items layout.